backend/requirements.txt
# ============================================
# RAG Lab - Backend Dependencies
# ============================================

# --------------------------------------------
# Web Framework
# --------------------------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# --------------------------------------------
# Google AI (Gemini + Embeddings)
# --------------------------------------------
google-generativeai==0.3.2
langchain-google-genai==0.0.6

# --------------------------------------------
# LangChain Core (versões compatíveis)
# --------------------------------------------
langchain==0.3.17
langchain-core==0.3.78
langchain-community==0.3.16

# --------------------------------------------
# LangGraph (Agentic RAG)
# --------------------------------------------
langgraph==0.2.70
langgraph-prebuilt==0.1.5

# --------------------------------------------
# Vector Database
# --------------------------------------------
pinecone-client==5.0.1
langchain-pinecone==0.0.1

# --------------------------------------------
# Reranking
# --------------------------------------------
cohere==4.47

# --------------------------------------------
# Graph Database (Neo4j - Opcional)
# --------------------------------------------
neo4j==5.16.0
langchain-neo4j==0.6.0

# --------------------------------------------
# Document Processing
# --------------------------------------------
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.2
unstructured==0.12.0

# --------------------------------------------
# Text Processing
# --------------------------------------------
tiktoken==0.5.2
sentence-transformers==2.3.1

# --------------------------------------------
# RAGAS (Evaluation)
# --------------------------------------------
ragas==0.1.1

# --------------------------------------------
# Utilities
# --------------------------------------------
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
tqdm==4.66.1

# --------------------------------------------
# HTTP & WebSocket
# --------------------------------------------
httpx==0.26.0
websockets==12.0

# --------------------------------------------
# Database (SQLAlchemy)
# --------------------------------------------
sqlalchemy==2.0.23
alembic==1.13.1

# --------------------------------------------
# Data Validation & Serialization
# --------------------------------------------
orjson==3.9.12

# --------------------------------------------
# Monitoring & Logging
# --------------------------------------------
loguru==0.7.2

# --------------------------------------------
# Development Tools
# --------------------------------------------
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# --------------------------------------------
# Optional: Advanced Features
# --------------------------------------------
# Redis (caching)
redis==5.0.1

# OpenTelemetry (observability)
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Prometheus (metrics)
prometheus-client==0.19.0
